+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/**/*
 .expo/*
+yarn-error.log
 npm-debug.*
 *.jks
 *.p8
