+++ constants/Colors.js
@@ -1,6 +1,4 @@
 export default {
-  // Morado Platanus
   primaryColor: '#2f2550',
-  // Amarillo Platanus
   secondColor: '#fabc13',
 };
