+++ screens/AuthScreen/styles.js
@@ -1,4 +1,3 @@
 import crossStyles from '../../styles/CrossStyles';
 
-const styles = crossStyles;
-export default styles;
+export default { ...crossStyles };
