+++ screens/HomeScreen/styles.js
@@ -1,4 +1,10 @@
+import { StyleSheet } from 'react-native';
 import crossStyles from '../../styles/CrossStyles';
 
-const styles = crossStyles;
-export default styles;
+const styles = StyleSheet.create({
+  saldoText: {
+    fontSize: 30,
+    padding: 20,
+  },
+});
+export default { ...crossStyles, ...styles };
