+++ screens/WelcomeScreen/WelcomeScreen.js
@@ -10,7 +10,7 @@ function WelcomeScreen(props) {
       <SocialIcon
         title='Sign In With Buda'
         button onPress={() => {
-          props.navigation.navigate({ routeName: 'Autentificacion' });
+          props.navigation.navigate({ routeName: 'Authentication' });
         }}
         type='medium'
       />
