+++ screens/WelcomeScreen/styles.js
@@ -1,11 +1,3 @@
-import { StyleSheet } from 'react-native';
 import crossStyles from '../../styles/CrossStyles';
 
-const styles = StyleSheet.create({
-  saldoText: {
-    fontSize: 30,
-    padding: 20,
-  },
-});
-
-export default { ...crossStyles, ...styles };
+export default { ...crossStyles };
