+++ store/sagas.js
@@ -13,7 +13,7 @@ export default function *rootSaga() {
           yield call(saga);
           break;
         } catch (e) {
-          console.log(e);
+          console.error(e);
         }
       }
     })),
